Internal: address review comments on design system import
- Move event const and payload type to @elementor/editor-canvas (shared between producer and listeners)
- Move Variable / StyleVariables types to @elementor/editor-styles for shared typing
- Type ImportedGlobalStylesPayload using StyleDefinitionsMap, StyleDefinitionID and StyleVariables
- Remove temporary trigger button injection (TriggerButton kept for the upcoming design system tab)
- Remove QA-only artificial import delay
