+++ packages/packages/core/editor-editing-panel/src/components/design-system-import/hooks/use-import-request.ts
@@ -1,3 +1,4 @@
+import { GLOBAL_STYLES_IMPORTED_EVENT, type ImportedGlobalStylesPayload } from '@elementor/editor-canvas';
 import { reloadCurrentDocument } from '@elementor/editor-documents';
 import { httpService } from '@elementor/http-client';
 import { useMutation } from '@elementor/query';
@@ -5,24 +6,13 @@ import { useMutation } from '@elementor/query';
 import { type ConflictStrategy } from '../types';
 
 const IMPORT_ENDPOINT = '/design-system/import';
-const GLOBAL_STYLES_IMPORTED_EVENT = 'elementor/global-styles/imported';
 
 export const IMPORT_DESIGN_SYSTEM_MUTATION_KEY = [ 'design-system-import' ] as const;
-
-// TEMP: artificial delay so the in-progress state is visible during manual QA. Remove before merging.
-const QA_DELAY_MS = 3000;
 
 type ImportRequestArgs = {
 	file: File;
 	conflictStrategy: ConflictStrategy;
 };
-
-export type ImportedGlobalStylesPayload = {
-	global_variables?: { data: Record< string, unknown > };
-	global_classes?: { items: Record< string, unknown >; order: string[] };
-};
-
-const wait = ( ms: number ): Promise< void > => new Promise( ( resolve: () => void ) => setTimeout( resolve, ms ) );
 
 export const useImportRequest = () => {
 	return useMutation( {
@@ -31,8 +21,6 @@ export const useImportRequest = () => {
 			const formData = new FormData();
 			formData.append( 'file', file );
 			formData.append( 'conflict_strategy', conflictStrategy );
-
-			await wait( QA_DELAY_MS );
 
 			const response = await httpService().post< ImportedGlobalStylesPayload >( IMPORT_ENDPOINT, formData, {
 				headers: { 'Content-Type': 'multipart/form-data' },
