--- packages/packages/core/editor-editing-panel/src/components/design-system-import/index.ts
@@ -1,1 +1,0 @@
-export { init } from './init';
