+++ packages/packages/core/editor-editing-panel/src/init.ts
@@ -2,7 +2,6 @@ import { injectIntoLogic } from '@elementor/editor';
 import { __registerPanel as registerPanel } from '@elementor/editor-panels';
 import { blockCommand } from '@elementor/editor-v1-adapters';
 
-import { init as initDesignSystemImport } from './components/design-system-import/init';
 import { EditingPanelHooks } from './components/editing-panel-hooks';
 import { init as initPromotionsSections } from './components/promotions/init';
 import { registerElementControls } from './controls-registry/element-controls/registry';
@@ -33,8 +32,6 @@ export function init() {
 	initResetStyleProps();
 
 	initPromotionsSections();
-
-	initDesignSystemImport();
 }
 
 const blockV1Panel = () => {
