+++ packages/packages/core/editor-global-classes/src/components/global-styles-import-listener.tsx
@@ -1,4 +1,5 @@
 import { useEffect } from 'react';
+import { GLOBAL_STYLES_IMPORTED_EVENT } from '@elementor/editor-canvas';
 import { __useDispatch as useDispatch } from '@elementor/store';
 
 import { apiClient } from '../api';
@@ -47,13 +48,10 @@ export function GlobalStylesImportListener() {
 				.catch( () => {} );
 		};
 
-		window.addEventListener( 'elementor/global-styles/imported', handleGlobalStylesImported as EventListener );
+		window.addEventListener( GLOBAL_STYLES_IMPORTED_EVENT, handleGlobalStylesImported as EventListener );
 
 		return () => {
-			window.removeEventListener(
-				'elementor/global-styles/imported',
-				handleGlobalStylesImported as EventListener
-			);
+			window.removeEventListener( GLOBAL_STYLES_IMPORTED_EVENT, handleGlobalStylesImported as EventListener );
 		};
 	}, [ dispatch ] );
