+++ packages/packages/core/editor-variables/src/components/global-styles-import-listener.tsx
@@ -1,4 +1,5 @@
 import { useEffect } from 'react';
+import { GLOBAL_STYLES_IMPORTED_EVENT } from '@elementor/editor-canvas';
 
 import { service } from '../service';
 import { styleVariablesRepository } from '../style-variables-repository';
@@ -19,13 +20,10 @@ export function GlobalStylesImportListener() {
 			service.load();
 		};
 
-		window.addEventListener( 'elementor/global-styles/imported', handleGlobalStylesImported as EventListener );
+		window.addEventListener( GLOBAL_STYLES_IMPORTED_EVENT, handleGlobalStylesImported as EventListener );
 
 		return () => {
-			window.removeEventListener(
-				'elementor/global-styles/imported',
-				handleGlobalStylesImported as EventListener
-			);
+			window.removeEventListener( GLOBAL_STYLES_IMPORTED_EVENT, handleGlobalStylesImported as EventListener );
 		};
 	}, [] );
