+++ packages/packages/core/editor-variables/src/types.ts
@@ -1,17 +1,8 @@
 import type * as React from 'react';
+import { type StyleVariables, type Variable } from '@elementor/editor-styles';
 import { type VirtualizedItem } from '@elementor/editor-ui';
 
-export type Variable = {
-	key?: string;
-	label: string;
-	value: string;
-	type: string;
-	deleted?: boolean;
-	deleted_at?: string;
-	sync_to_v3?: boolean;
-};
-
-export type StyleVariables = Record< string, Variable >;
+export type { StyleVariables, Variable };
 
 export type ExtendedVirtualizedItem = VirtualizedItem< 'item', string > & {
 	icon: React.ReactNode;
