+++ packages/packages/libs/editor-styles/src/types.ts
@@ -47,3 +47,15 @@ export type StyleDefinition = {
 };
 
 export type StyleDefinitionsMap = Record< StyleDefinition[ 'id' ], StyleDefinition >;
+
+export type Variable = {
+	key?: string;
+	label: string;
+	value: string;
+	type: string;
+	deleted?: boolean;
+	deleted_at?: string;
+	sync_to_v3?: boolean;
+};
+
+export type StyleVariables = Record< string, Variable >;
